Completely restructure how we connect to the database(s).
The previous design meant that lots of pieces of code had to copy/paste logic
about deciding whether to connect to a "test" or "real" database. We now
centralize this logic in a few key functions, which in turn use a restructured
configuration file to decide which database to connect to and whether to treat
it as a "production" or "testing" configuration.

We also add functions that automatically generate and use an
argparse.ArgumentParser to mean that command-line programs can have a unified
interface for choosing which database to use.

The new configuration file allows an arbitrary list of database connections,
one of which is marked as the default. The testing code looks for the
existence of a database called "testing".

I've also altered the test_mc.py test rig to not perform any operations on the
"real" database. I don't want to run the risk of someone running tests that
alter the production database, no matter how hard we try to ensure that any
changes are kept within transactions, etc!