Move most database operations into a custom Session class, rather than a custom DB.
The actual operations on the database all need a Session object, so it's the
natural place to implement our features.

We also turn the Session object into a Python "context manager", which lets us
use Sessions in "with" statements that help us be sure that the session either
gets fully committed to the database, or rolled back.