scripts/mc_server_status_daemon.py: add server status reporter
Various HERA machines are supposed to periodically report
their status to the M&C system. Most of the information can
be gathered fairly straightforwardly, but we do want to average
load values over time to beat down the noise somewhat. As best
I (PKGW) can tell, the most straightforward way to do this is
to write a small daemon script that maintains and M&C database
connection and does the averaging. Lo and behold, here we are.

We also fix a mistake in mc_session.py.