scripts/mc_megaraid_monitor.py: script to monitor RAID statuses
Part of M&C's job is to check on the status of our RAID arrays. This script
uses `storcli64` to assess the RAID status and log it to the database. There
is all sorts of information that we *could* log, but hopefully this analysis
covers the most important things by watching the event log for errors.